Add write_data_datagram to circular buffer
Copies header and data into a new char array and sends it
in one go as a large datagram. Not sure if this will affect
memory/speed enough to matter.

Need to clean things up as it was a quick and dirty test. Can
remove write_data now that we have this method.